+++ internal/cmd/artifacts/list.go
@@ -1,11 +1,14 @@
 package artifacts
 
 import (
+	"encoding/json"
 	"errors"
 	"fmt"
+	"os"
 
 	"github.com/jedib0t/go-pretty/v6/table"
 	"github.com/jedib0t/go-pretty/v6/text"
+	"github.com/saucelabs/saucectl/internal/artifacts"
 	cmds "github.com/saucelabs/saucectl/internal/cmd"
 	"github.com/saucelabs/saucectl/internal/segment"
 	"github.com/saucelabs/saucectl/internal/usage"
@@ -96,5 +99,46 @@ func list(jobID, outputFormat string) error {
 		return fmt.Errorf("failed to get artifacts list: %w", err)
 	}
 
-	return renderResults(lst, outputFormat)
+	switch outputFormat {
+	case "json":
+		if err := renderJSON(lst); err != nil {
+			return fmt.Errorf("failed to render output: %w", err)
+		}
+	case "text":
+		renderTable(lst)
+	default:
+		return errors.New("unknown output format")
+	}
+
+	return nil
+}
+
+func renderTable(lst artifacts.List) {
+	if len(lst.Items) == 0 {
+		println("No artifacts for this job.")
+		return
+	}
+
+	t := table.NewWriter()
+	t.SetStyle(defaultTableStyle)
+	t.SuppressEmptyColumns()
+
+	t.AppendHeader(table.Row{"Items"})
+	t.SetColumnConfigs([]table.ColumnConfig{
+		{
+			Name: "Items",
+		},
+	})
+
+	for _, item := range lst.Items {
+		// the order of values must match the order of the header
+		t.AppendRow(table.Row{item})
+	}
+	t.SuppressEmptyColumns()
+
+	println(t.Render())
+}
+
+func renderJSON(val any) error {
+	return json.NewEncoder(os.Stdout).Encode(val)
 }
